Ki's Adventures
This game is written in Python 2.7 and utilizes curses (cursor optimization module), time, random. The game is about a character called Ki who is in search of a sword. Ki has to pass 20 levels before reaching the sword. He has five HP which can be lost over time when he is hit by a bullet. The controls are: Menu screen, up key, down key, and the "Return" key. Game: "Return" key (to pause), up key (to jump), left/right keys (to move left and right), and the "q" key to exit (you can only exit after pausing). 

NOTICE, this game still has bugs (Only one bug that I know of... For now...), For a example in the start of the game (level 1) you might glitch through the boulder.